backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.5
pydantic-settings==2.7.1
python-dotenv==1.0.1
motor==3.6.1
pymongo==4.9.2
cloudinary==1.44.2
google-generativeai==0.8.3
httpx==0.27.2
cachetools==5.5.0
PyMuPDF==1.25.1
pdfplumber==0.11.5
beautifulsoup4==4.12.3
sentence-transformers==3.3.1
faiss-cpu==1.9.0.post1
numpy==1.26.4
langchain==0.3.14
pytest==8.3.4
pytest-asyncio==0.25.2
respx==0.21.1
mongomock-motor==0.0.35
